finishing the posts routes

+++ src/services/postService.js
@@ -35,29 +35,29 @@ export async function getPostsOfFollowing(user_id) {
     return rows[0]
 }
 
-export async function deletePost(id) {
+export async function deletePost(post_id, user_id) {
     const conn = await db.connect()
-    const sql = 'DELETE FROM Posts WHERE id =?'
-    const data = [id]
+    const sql = 'DELETE FROM Posts WHERE id =? AND user_id =?'
+    const data = [post_id, user_id]
     await conn.query(sql, data)
     await conn.end()
 }
 
-export async function updatePost(id, caption, image) {
+export async function updatePost(post_id, user_id, caption, image) {
     const conn = await db.connect()
-    const sql = 'UPDATE Posts SET caption =?, image =? WHERE id =?'
-    const data = [caption, image, id]
+    const sql = 'UPDATE Posts SET caption =?, image =?, updated =true WHERE id =? AND user_id =?'
+    const data = [caption, image, post_id, user_id]
     await conn.query(sql, data)
     await conn.end()
 }
 
-export async function getPostsByAuthor(nickname) {
+export async function getPostsOfUser(user_id) {
     const conn = await db.connect()
-    const sql = 'SELECT * FROM PostView WHERE nickname =?'
-    const data = [nickname]
+    const sql = 'SELECT * FROM PostView WHERE user_id =?'
+    const data = [user_id]
     const rows = await conn.query(sql, data)
     conn.end()
-    return rows
+    return rows[0]
 }
 
 export async function likePost(user_id, post_id) {
@@ -94,4 +94,16 @@ export async function getNumOfLikes(id){
     return row[0]
 }
 
-export default { createPost, getPost, getAllPosts, updatePost, getPostsByAuthor, likePost, unlikePost, checkUserLikedPost, getNumOfLikes, getPostsOfFollowing }
+export default { 
+    createPost, 
+    getPost, 
+    getAllPosts, 
+    updatePost, 
+    deletePost,
+    getPostsOfUser, 
+    likePost, 
+    unlikePost, 
+    checkUserLikedPost, 
+    getNumOfLikes, 
+    getPostsOfFollowing 
+}
